4.x: Make HttpRouting an interface
